feat: add optimize folder and move evaluator
